rename sidebarViewProvider to sidebarViewController
